auth-service/requirements.txt
fastapi
uvicorn[standard]
python-jose[cryptography]
passlib[argon2]
email-validator
python-dotenv
argon2-cffi
SQLAlchemy